requirements.txt
fastapi
uvicorn
qrcode[pil]
pillow
requests
psycopg2-binary
python-dotenv